+++ test/functional/android/webdriver/w3c_actions_test.rb
@@ -156,7 +156,6 @@ class AppiumLibCoreTest
                 ::Selenium::WebDriver::Error::InvalidArgumentError].include? error.class
       end
 
-      # NOTE: Works with Espresso Driver
       def test_multiple_actions
         skip_as_appium_version '1.8.0'